infrastructure/requirements.txt
aws-cdk-lib==2.95.0
constructs==10.3.0
